messaging_app/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-filter==24.3
mysqlclient==2.2.4
pytest==8.2.1
pytest-django==4.8.0
pytest-cov==5.0.0
flake8==7.1.0
gunicorn==21.2.0
